pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glacis-autoredteam"
version = "0.1.1"
description = "Closed-loop adversarial red-teaming and hardening for AI systems"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    {name = "Glacis", email = "eng@glacis.io"},
]
keywords = ["ai-safety", "red-teaming", "llm", "security", "alignment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0,<3.0"]
anthropic = ["anthropic>=0.30,<1.0"]
google = ["google-generativeai>=0.8,<1.0"]
cloud = ["boto3>=1.34,<2.0"]
training = [
    "torch>=2.0",
    "transformers>=4.40",
    "peft>=0.10",
    "accelerate>=0.28",
    "datasets>=2.18",
]
all = [
    "glacis-autoredteam[openai]",
    "glacis-autoredteam[anthropic]",
    "glacis-autoredteam[google]",
    "glacis-autoredteam[cloud]",
    "glacis-autoredteam[training]",
]

[project.urls]
Homepage = "https://github.com/glacis-io/auto-redteam"
Documentation = "https://github.com/glacis-io/auto-redteam"
Repository = "https://github.com/glacis-io/auto-redteam"

[tool.setuptools]
py-modules = [
    "attack",
    "attestation",
    "autoharden",
    "campaign",
    "campaign_runner",
    "cli",
    "conversation",
    "defend",
    "immune",
    "prepare",
    "run",
    "scoring",
    "scoring_v2",
    "stealth",
    "training_data",
    "trajectory_engine",
]

[tool.setuptools.packages.find]
include = ["autoredteam*", "providers*", "attack_packs*", "training*", "reporting*", "models*"]
exclude = ["tests*", "training_data*", "results*", "demo*"]
